modified the CNN software architecture
