+++ gilbreth_msgs/msg/ObjectType.msg
@@ -1,4 +1,4 @@
 std_msgs/Header header    # Message Header    
 time detection_time       # The time when the object was detected
-std_msgs/Int32MultiArray voxel   # Voxel data of object
+int8 type   # Object type
 sensor_msgs/PointCloud2 pcd # Point cloud data
