+++ gilbreth_msgs/package.xml
@@ -16,6 +16,7 @@
   <depend>message_generation</depend>
   <depend>message_runtime</depend>
   <depend>moveit_msgs</depend>
+  <depend>sensor_msgs</depend>
 
 
 </package>
