+++ gilbreth_perception/README.md
@@ -1,9 +1,7 @@
-1.Save gilbreth_perception folder under your-catkin-workspace/src/gilbreth.
-2.Under your-catkin-workspace run command "catkin_make".
-3.Run command "roslaunch gilbreth_perception gilbreth_perception.launch".
+For using Correspondence Grouping method in object recognition:
+1.Run command "roslaunch gilbreth_perception gilbreth_perception.launch cnn:=false".
 
-For using convolutional neural network (instead of the correspondence grouping) in the recognition node:
-
+For using convolutional neural network method in object recognition:
 1. Make sure your computer has Nvidia GPU;
 2. Install voxnet at https://github.com/dimatura/voxnet;
 - Install all dependent packages as instructed on the webpage
@@ -12,5 +10,5 @@ For using convolutional neural network (instead of the correspondence grouping) 
 - Verify voxnet has been installed successfully by ‘import voxnet’ in python, make sure no error occurs
 - If you have error like ‘CudaNdarrayType only supports dtype float32 for now’ , solution is here:
   https://github.com/dimatura/voxnet/issues/5
-3. Save your neural network weights into src/gilbreth/gilbreth_perception/config/
-4. Run command "roslaunch gilbreth_perception gilbreth_perception_cnn.launch"
+3. Save your neural network weights into src/gilbreth/gilbreth_perception/config/weights.npz
+4. Run command "roslaunch gilbreth_perception gilbreth_perception.launch"
