+++ gilbreth_perception/package.xml
@@ -23,7 +23,4 @@
   <run_depend>sensor_msgs</run_depend>
   <build_depend>rospy</build_depend>
   <run_depend>rospy</run_depend>
-  <build_depend>python-numpy</build_depend>
-  <run_depend>python-numpy</run_depend>
-
 </package>
